refac: moved components
