+++ lib/components/FateLabel/FateLabel.tsx
